In progress Google Sign In.
Switched to session for majority of checking authentication. Much faster and snappy! But we still make use of loadUser() as we will not be attaching the entire user object to the browser session (security reason). So it's still useful in some cases.

Going to implement edge case features now, to begin supporting google sign in for real.