Added For Experts, For Advice Seekers, Meet the Team page, Support page.
Experimented with updating and caching payment intent, but is defunct for now.

Lots of frontend additions

updated user and expertise models to reflect Stripe Customer saving feature.

Tried to implement using saved payment method, but also canned it for now, as would require more than i thought.

Last step is to add live demos, and screenshots on the for experts and for advice seekers pages and expert dashboard.